requirements.txt
# Core ML/RL dependencies
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# SUMO
traci>=1.19.0

# Training & utilities
tensorboard>=2.15.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Optional
pillow>=10.0.0
